# cmd/api/docs/swagger.yaml
definitions:
...
  models.MakeReadReq:
    properties:
      message_ids:
        items:
          type: string
        type: array
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      statusCode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /chat:
    get:
      consumes:
      - application/json
      description: Get all chats associated with the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: successfully showing chats of the user
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: unauthorised
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user's chats
      tags:
      - Chat
  /chat/{chatId}/message:
    get:
      consumes:
      - application/json
      description: Get all messages in a chat based on the provided chatId.
      parameters:
      - description: chatID
        format: objectId
        in: path
        name: chatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully showing messages in the given chatId
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: string converrsion failed
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get messages in a chat
      tags:
      - Chat
  /chat/message/read:
    post:
      consumes:
      - application/json
      description: Marks specified messages as read by the user.
      operationId: makeMessageRead
      parameters:
      - description: Message IDs to mark as read
        in: body
        name: messageIds
        required: true
        schema:
          $ref: '#/definitions/models.MakeReadReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully marked messages as read
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Mark messages as read
      tags:
      - Chat
...
  /ws/{chatId}:
    get:
      consumes:
      - application/json
      description: Handles WebSocket connection for chat and processes incoming messages
      operationId: chat-websocket
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Handle WebSocket connection for chat
      tags:
      - Chat
swagger: "2.0"